# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.ProjectCapacityResponse:
    properties:
      available_slots:
        type: integer
      has_capacity:
        type: boolean
      id:
        type: string
      max_workers:
        type: integer
      name:
        type: string
      queue_size:
        type: integer
      running:
        type: integer
    type: object
  handler.ProjectResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: abc123
        type: string
      name:
        example: My Project
        type: string
      path:
        example: /Users/user/projects/myproject
        type: string
    type: object
  handler.ProjectsListResponse:
    properties:
      projects:
        items:
          $ref: '#/definitions/handler.ProjectResponse'
        type: array
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@openvibely.dev
    name: API Support
    url: https://github.com/openvibely/openvibely
  description: |-
    REST API for OpenVibely - AI-powered task scheduling and management
    This API provides endpoints for managing projects, tasks, and chat interactions with AI agents.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OpenVibely API
  version: "1.0"
paths:
...
  /api/capacity/projects:
    get:
      description: Returns worker capacity and queue information for each project.
      produces:
      - application/json
      responses:
        "200":
          description: Per-project capacity information
          schema:
            items:
              $ref: '#/definitions/handler.ProjectCapacityResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get project worker capacities
      tags:
      - capacity
  /api/capacity/projects/{projectId}:
    get:
      description: Returns worker capacity and queue information for a specific project.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project capacity information
          schema:
            $ref: '#/definitions/handler.ProjectCapacityResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get project worker capacity
      tags:
      - capacity
...
  /api/projects:
    get:
      description: Get a list of all available projects with their IDs, names, paths,
        and creation timestamps
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved projects
          schema:
            $ref: '#/definitions/handler.ProjectsListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List all projects
      tags:
      - projects
...
schemes:
- http
- https
swagger: "2.0"
...